Remove `PackageRootedDependencyMap` in favor of `SymbolMap`.
[ci skip-rust]

[ci skip-build-wheels]
