+++ src/python/pants/backend/experimental/java/register.py
@@ -17,7 +17,7 @@ from pants.backend.java.target_types import rules as target_types_rules
 from pants.backend.java.test import junit
 from pants.jvm import classpath, jdk_rules
 from pants.jvm import util_rules as jvm_util_rules
-from pants.jvm.dependency_inference import package_mapper
+from pants.jvm.dependency_inference import symbol_mapper
 from pants.jvm.goals import coursier
 from pants.jvm.resolve import coursier_fetch, coursier_setup
 from pants.jvm.target_types import JvmArtifact, JvmDependencyLockfile
@@ -48,7 +48,7 @@ def rules():
         *import_parser.rules(),
         *java_parser.rules(),
         *java_parser_launcher.rules(),
-        *package_mapper.rules(),
+        *symbol_mapper.rules(),
         *dependency_inference_rules.rules(),
         *tailor.rules(),
         *jvm_util_rules.rules(),
