+++ src/python/pants/backend/java/dependency_inference/symbol_mapper.py
@@ -7,15 +7,12 @@ import logging
 
 from pants.backend.java.dependency_inference.types import JavaSourceDependencyAnalysis
 from pants.backend.java.target_types import JavaSourceField
-from pants.core.util_rules.source_files import SourceFiles, SourceFilesRequest
+from pants.core.util_rules.source_files import SourceFilesRequest
 from pants.engine.rules import Get, MultiGet, collect_rules, rule
 from pants.engine.target import AllTargets, Targets
 from pants.engine.unions import UnionRule
-from pants.jvm.dependency_inference.package_mapper import (
-    FirstPartyMappingImpl,
-    FirstPartyMappingRequest,
-)
-from pants.jvm.dependency_inference.package_prefix_tree import PackageRootedDependencyMap
+from pants.jvm.dependency_inference import symbol_mapper
+from pants.jvm.dependency_inference.symbol_mapper import FirstPartyMappingRequest, SymbolMap
 from pants.util.logging import LogLevel
 
 logger = logging.getLogger(__name__)
@@ -37,26 +34,24 @@ class FirstPartyJavaTargetsMappingRequest(FirstPartyMappingRequest):
 @rule(desc="Map all first party Java targets to their packages", level=LogLevel.DEBUG)
 async def map_first_party_java_targets_to_symbols(
     _: FirstPartyJavaTargetsMappingRequest, java_targets: AllJavaTargets
-) -> FirstPartyMappingImpl:
-    source_files = await MultiGet(
-        Get(SourceFiles, SourceFilesRequest([target[JavaSourceField]])) for target in java_targets
-    )
+) -> SymbolMap:
     source_analysis = await MultiGet(
-        Get(JavaSourceDependencyAnalysis, SourceFiles, source_files)
-        for source_files in source_files
+        Get(JavaSourceDependencyAnalysis, SourceFilesRequest([target[JavaSourceField]]))
+        for target in java_targets
     )
     address_and_analysis = zip([t.address for t in java_targets], source_analysis)
 
-    dep_map = PackageRootedDependencyMap()
+    dep_map = SymbolMap()
     for address, analysis in address_and_analysis:
         for top_level_type in analysis.top_level_types:
-            dep_map.add_top_level_type(top_level_type, address=address)
+            dep_map.add_symbol(top_level_type, address=address)
 
-    return FirstPartyMappingImpl(package_rooted_dependency_map=dep_map)
+    return dep_map
 
 
 def rules():
     return (
         *collect_rules(),
+        *symbol_mapper.rules(),
         UnionRule(FirstPartyMappingRequest, FirstPartyJavaTargetsMappingRequest),
     )
