+++ src/python/pants/jvm/dependency_inference/BUILD
@@ -2,7 +2,3 @@
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
 python_sources()
-
-python_tests(
-    name="tests",
-)
